requirements.txt
pytz==2023.3
python-telegram-bot==20.6
httpx==0.27.0
python-dotenv==1.0.0